delete queues
