+++ services/core/service/peer/consumer/sender_queue.go
@@ -1,4 +1,4 @@
-package amqp
+package consumer
 
 import (
 	"broker/core/config"
