+++ services/core/service/peer/publisher/receiver_dispatcher.go
@@ -1,4 +1,4 @@
-package amqp
+package publisher
 
 import (
 	"broker/pkg/logger"
