+++ services/core/service/peer/publisher/receiver_queue.go
@@ -1,4 +1,4 @@
-package amqp
+package publisher
 
 import (
 	"broker/core/config"
